Refactor cli/main.py: extract subcommand bodies into _register_*.py modules
main.py had grown to 1,386 lines with seven Typer commands defined inline
(scan: 237, init: 326, doctor: 205) and ~350 lines of shared helpers at
the bottom. The file was hard to navigate and edits regularly conflicted.

Extract each inline command into its own private module
(cli/_register_<name>.py) exposing register(app: typer.Typer) -> None.
Move the 12 shared helpers (parsers, diagnostics, multi-scan loop,
summary printing) into cli/_helpers.py. main.py shrinks to an 89-line
registration shell. baseline_app now lives inside _register_baseline.

Registration order in main.py preserves byte-identical --help output:
external commands first, then the seven _register_* calls (baseline
last to keep the baseline/fixture/scenario subgroup order), then
add_typer for fixture and scenario.

No behavior change. tests/test_cli.py updated to import the two private
helpers (_safe_output_name, _tool_surface_diff_has_changes) from
_helpers instead of main.

Verified: ruff clean, 1049/1049 pytest, --help byte-identical across
all 8 surfaces (root + scan + list-checks + contract + explain + init
+ doctor + baseline + baseline save).
